requirements.txt
flask
flask-sqlalchemy
flask-cors
mysql-connector-python
python-dotenv
pytest
pytest-flask
